add: ReqNats. chore: reorganize code.
